pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "project_name"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pyvis >= 0.3.2",
    "networkx",
    "requests>=2.32.3",
    "pywikibot>=10.1.0",
    "pandas>=2.2.0",
    "mwparserfromhell>=0.6.6",
    "google-api-core>=2.11.0",
    "google-api-python-client>=2.172.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "googleapis-common-protos>=1.70.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
